Adicionando Exceptions e atualizando README

+++ miltankbank/src/main/java/br/com/miltankbank/service/executor/exclui/ExecutorExcluiDespesa.java
@@ -1,7 +1,11 @@
 package br.com.miltankbank.service.executor.exclui;
 
+import java.util.Optional;
+
 import org.springframework.stereotype.Service;
 
+import br.com.miltankbank.exceptions.despesa.DespesaExcluidaException;
+import br.com.miltankbank.model.entity.Despesa;
 import br.com.miltankbank.model.repository.DespesaRepository;
 import br.com.miltankbank.service.acao.AcaoExcluiDespesa;
 
@@ -16,7 +20,11 @@ public class ExecutorExcluiDespesa implements AcaoExcluiDespesa {
 
     @Override
     public void executa(Long idDespesa) {
-        despesaRepository.deleteById(idDespesa);
+        Optional<Despesa> optDespesa = despesaRepository.findById(idDespesa);
+        if (optDespesa.isPresent()) {
+            despesaRepository.deleteById(idDespesa);
+        }
+        throw new DespesaExcluidaException("A despesa de id: "  + idDespesa + " já foi excluída anteriormente");
     }
 
 }
