+++ miltankbank/src/main/java/br/com/miltankbank/service/executor/exclui/ExecutorExcluiReceita.java
@@ -1,7 +1,11 @@
 package br.com.miltankbank.service.executor.exclui;
 
+import java.util.Optional;
+
 import org.springframework.stereotype.Service;
 
+import br.com.miltankbank.exceptions.receita.ReceitaExcluidaException;
+import br.com.miltankbank.model.entity.Receita;
 import br.com.miltankbank.model.repository.ReceitaRepository;
 import br.com.miltankbank.service.acao.AcaoExcluiReceita;
 
@@ -16,6 +20,10 @@ public class ExecutorExcluiReceita implements AcaoExcluiReceita {
 
     @Override
     public void executa(Long idReceita) {
-        receitaRepository.deleteById(idReceita);
+        Optional<Receita> optReceita = receitaRepository.findById(idReceita);
+        if (optReceita.isPresent()) {
+            receitaRepository.deleteById(idReceita);    
+        }
+        throw new ReceitaExcluidaException("A receita de id: "  + idReceita + " já foi excluída anteriormente");
     }
 }
